% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bild_ps1_from_xlsx.R
\name{build_ps1_from_xlsx}
\alias{build_ps1_from_xlsx}
\title{Build ps1 files}
\usage{
build_ps1_from_xlsx(file = file.choose(), pwd)
}
\arguments{
\item{file}{(chr, default interactive selection windows) file path to
the Excel file reporting the users to create (see details).}

\item{pwd}{(chr, optional) initial password shared by every account in
the batch. When omitted a strong random password is generated for the
run. It is embedded in the generated \code{.ps1}; users are always forced
to change it at first sign-in.}
}
\description{
From a correctly formatted Excel file, generate the Microsoft Graph
PowerShell (\code{New-MgUser}/\code{Remove-MgUser}) scripts to create and delete
the users in bulk on Microsoft Entra ID, plus the CSV to add them to a
group from the Azure portal.
}
\details{
The Excel file must have 9 column, named exactly (case sensitive) as:
\verb{Nome, Cognome, Email, Prj1_ID, Prj1_role, Prj1_DAG, Prj2_ID, Prj2_role, Prj2_DAG}.

Moreover:
\itemize{
\item no more than 2 projects can be added at the creation time using
this script.
\item At least information (all the three!) for one project must be
provided for EDCxx servers
\item For students on MSTxx servers no information should be provided
about projects (i.e. the first three column should contain data
only!)
\item For template including a projects all the role and DAG for each
of them must be provided!
}
}
\examples{
\dontrun{
  build_ps1_from_xlsx()
}

}
