% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provisioning_scope.R
\name{scope_filled}
\alias{scope_filled}
\title{Is this cell filled in}
\usage{
scope_filled(values)
}
\arguments{
\item{values}{A vector from a register column.}
}
\value{
A logical vector.
}
\description{
Is this cell filled in
}
\keyword{internal}
